[otbn,rtl] Fix start_stop_control done_o signal when SecWipeEn=0
This single-cycle pulse is supposed to happen at the end of the secure
wipe. When SecWipeEn=0, we have a single-cycle "secure wipe" that just
jumps straight to OtbnStartStopSecureWipeComplete. But the done_o
signal didn't match: fix that.

There's also a corresponding DV change needed to get the status signal
timing right.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>